requirements.txt
Pillow==9.0.1
pyfiglet==1.0.2
python-telegram-bot==21.7
transliterate==1.10.2
setproctitle
python-telegram-bot[job-queue]
